requirements.txt
Pillow
selenium
requests